Small fixes, there is a bug with memory allocation according to the Task Manager the memory used by the program increases by .2 MB every few seconds (not freeing somewhere)

+++ code/engine_file_formats_.h
@@ -380,8 +380,8 @@ struct kesa_tileset
     u32 TileWidth;
     u32 TileHeight;
 
-    u32 TileOffsetsX[512];
-    u32 TileOffsetsY[512];
+    u8 TileOffsetsX[512];
+    u8 TileOffsetsY[512];
 };
 
 struct kesa_sound
